+++ Exception/LoadConfigurationFailedException.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 /*
  * This file is part of the SwaggerResolverBundle package.
  *
- * (c) Viktor Linkin <adrenalinkin@gmail.com>
+ * (c) MarfaTech <https://marfa-tech.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -18,15 +18,12 @@ use RuntimeException;
 use function sprintf;
 
 /**
- * @author Viktor Linkin <adrenalinkin@gmail.com>
+ * @author MarfaTech <https://marfa-tech.com>
  */
-class DefinitionNotFoundException extends RuntimeException
+class LoadConfigurationFailedException extends RuntimeException
 {
-    /**
-     * @param string $className
-     */
-    public function __construct(string $className)
+    public function __construct(string $message)
     {
-        parent::__construct(sprintf('Swagger definition "%s" was not found', $className));
+        parent::__construct(sprintf('Load Open API configuration is failed: "%s"', $message));
     }
 }