+++ Validator/ArrayMinItemsValidator.php
@@ -13,7 +13,8 @@ declare(strict_types=1);
 
 namespace Linkin\Bundle\SwaggerResolverBundle\Validator;
 
-use EXSyst\Component\Swagger\Schema;
+use OpenApi\Annotations\Property;
+use OpenApi\Generator;
 use Symfony\Component\OptionsResolver\Exception\InvalidOptionsException;
 
 use function count;
@@ -27,24 +28,28 @@ class ArrayMinItemsValidator extends AbstractArrayValidator
     /**
      * {@inheritdoc}
      */
-    public function supports(Schema $property, array $context = []): bool
+    public function supports(Property $property): bool
     {
-        return parent::supports($property, $context) && null !== $property->getMinItems();
+        return parent::supports($property) && !Generator::isDefault($property->minItems);
     }
 
     /**
      * {@inheritdoc}
      */
-    public function validate(Schema $property, string $propertyName, $value): void
+    public function validate(Property $property, $value): void
     {
-        $value = $this->convertValueToArray($propertyName, $value, $property->getCollectionFormat());
+        if ($value === null) {
+            return;
+        }
 
-        if (count($value) < $property->getMinItems()) {
-            throw new InvalidOptionsException(sprintf(
-                'Property "%s" should have %s items or more',
-                $propertyName,
-                $property->getMinItems()
-            ));
+        $propertyName = $property->property;
+
+        $value = $this->convertValueToArray($propertyName, $value, $property->collectionFormat);
+
+        if (count($value) < $property->minItems) {
+            $message = sprintf('Property "%s" should have %s items or more', $propertyName, $property->minItems);
+
+            throw new InvalidOptionsException($message);
         }
     }
 }