+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 ## [Unreleased]
 ### Added
+- Support OpenApi 3.
+### Removed
+- [BC] Support Swagger 2.
+- [BC] Set `nullable` type if property isn't required.
+
+## [0.5.1] - 2022-06-15
+### Fixed
+- Set `nullable` type if property isn't required.
+- Forward all parent property meta while describe model.
+
+## [0.5.0] - 2022-05-17
+### Added
 - Accept and read swagger extension `x-nullable`.
 - Accept property annotation constraints.
 - Swagger extension `x-class` has full class name of body model.